+++ packages/aikit-ui/src/_ai-elements/message.tsx
@@ -2,7 +2,7 @@ import {
   Avatar,
   AvatarFallback,
   AvatarImage,
-} from '@hebo/ui/_shadcn/ui/avatar';
+} from '../_shadcn/ui/avatar';
 import type { ComponentProps, HTMLAttributes } from 'react';
 import { cn } from '@hebo/aikit-ui/lib/utils';
 import type { UIMessage } from 'ai';
