+++ packages/aikit-ui/src/_ai-elements/prompt-input.tsx
@@ -9,10 +9,10 @@ import {
   SelectItem,
   SelectTrigger,
   SelectValue,
-} from "@hebo/aikit-ui/_shadcn/select";
-import { Textarea } from "@hebo/aikit-ui/_shadcn/textarea";
+} from "@hebo/aikit-ui/_shadcn/ui/select";
+import { Textarea } from "@hebo/aikit-ui/_shadcn/ui/textarea";
 import { cn } from "@hebo/aikit-ui/lib/utils";
-import { Button } from "@hebo/ui/_shadcn/ui/button";
+import { Button } from "../_shadcn/ui/button";
 
 import type { ChatStatus } from "ai";
 import type {
