+++ packages/aikit-ui/src/_shadcn/ui/avatar.tsx
