+++ packages/aikit-ui/src/_shadcn/ui/sidebar.tsx
@@ -7,23 +7,23 @@ import { PanelLeftIcon } from "lucide-react"
 
 import { useIsMobile } from "@hebo/ui/hooks/use-mobile"
 import { cn } from "@hebo/ui/lib/utils"
-import { Button } from "@hebo/ui/_shadcn/ui/button"
-import { Input } from "@hebo/ui/_shadcn/ui/input"
-import { Separator } from "@hebo/ui/_shadcn/ui/separator"
+import { Button } from "./button"
+import { Input } from "./input"
+import { Separator } from "./separator"
 import {
   Sheet,
   SheetContent,
   SheetDescription,
   SheetHeader,
   SheetTitle,
-} from "@hebo/ui/_shadcn/ui/sheet"
-import { Skeleton } from "@hebo/ui/_shadcn/ui/skeleton"
+} from "./sheet"
+import { Skeleton } from "./skeleton"
 import {
   Tooltip,
   TooltipContent,
   TooltipProvider,
   TooltipTrigger,
-} from "@hebo/ui/_shadcn/ui/tooltip"
+} from "./tooltip"
 
 const SIDEBAR_COOKIE_NAME = "sidebar_state"
 const SIDEBAR_COOKIE_MAX_AGE = 60 * 60 * 24 * 7
