+++ packages/ui/package.json
@@ -11,6 +11,7 @@
     "clean": "rimraf .turbo node_modules"
   },
   "dependencies": {
+    "@hebo/aikit-ui": "workspace:*",
     "ai": "^5.0.10",
     "ai-elements": "^1.0.3",
     "class-variance-authority": "^0.7.1",
