+++ packages/ui/src/components/DropdownMenu.tsx
@@ -14,4 +14,4 @@ export {
   DropdownMenuSub,
   DropdownMenuSubTrigger,
   DropdownMenuSubContent,
-} from "@hebo/ui/_shadcn/ui/dropdown-menu";
+} from "@hebo/aikit-ui/src/_shadcn/ui/dropdown-menu";
